requirements.txt
ConfigArgParse==1.0
configparser==4.0.2
matplotlib==3.3.1
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
proglog==0.1.9
pyrender==0.1.33
PyYAML==5.3
scipy==1.4.1
smplx==0.1.13
torch==1.8.1
torchgeometry==0.1.2
tqdm==4.42.0
trimesh==3.5.16
joblib==0.16.0


